sending recovery link

+++ src/main/java/com/workflow/workflow/user/UserSessionRepository.java
@@ -5,7 +5,7 @@ import java.util.Optional;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface UserSessionRepository extends CrudRepository<UserSession, Integer> {
+public interface UserSessionRepository extends CrudRepository<UserSession, Long> {
     Optional<UserSession> findBySession(String session);
     List<UserSession> findByUser(User user);
 }
